Split yacoin-cli and yacoind
