+++ .github/workflows/yacoinbuildmultiplatform.yml
@@ -27,6 +27,10 @@ jobs:
         with:
           name: yacoin-qt-ubuntu-16.04
           path: /home/runner/work/yacoin/yacoin/release/yacoin-qt
+      - uses: actions/upload-artifact@v4
+        with:
+          name: yacoin-cli-ubuntu-16.04
+          path: /home/runner/work/yacoin/yacoin/release/yacoin-cli
   build-ubuntu-1604-functional-test:
     runs-on: ubuntu-latest  
     steps:
@@ -56,6 +60,10 @@ jobs:
         with:
           name: yacoin-qt-ubuntu-18.04-qt4
           path: /home/runner/work/yacoin/yacoin/release/yacoin-qt
+      - uses: actions/upload-artifact@v4
+        with:
+          name: yacoin-cli-ubuntu-18.04
+          path: /home/runner/work/yacoin/yacoin/release/yacoin-cli
   build-ubuntu-1804-qt5:
     runs-on: ubuntu-latest  
     steps:
