--- src/bitcoin-cli.h
@@ -1,12 +1,0 @@
-// Copyright (c) 2009-2010 Satoshi Nakamoto
-// Copyright (c) 2009-2016 The Bitcoin Core developers
-// Copyright (c) 2017-2025 The Yacoin Core developers
-// Distributed under the MIT software license, see the accompanying
-// file COPYING or http://www.opensource.org/licenses/mit-license.php.
-
-#ifndef YACOIN_CLI_H
-#define YACOIN_CLI_H
-
-int CommandLineRPC(int argc, char *argv[]);
-
-#endif // YACOIN_CLI_H
